rename some folders/files for better clarity
no functionality changes
