+++ src/routes.ts
@@ -1,7 +1,7 @@
 import { PublicKey } from "@solana/web3.js";
 import { Router } from "express";
 import { getKnex } from "./database";
-import { TBCAccount, TBCAccountBalance } from "./types/tbc_accounts";
+import { TBCAccount, TBCAccountBalance } from "./knex-types/tbc_accounts";
 
 const routes = Router();
